+++ Pandora/Pattern/Functor.hs
@@ -11,7 +11,6 @@ import Pandora.Pattern.Functor.Bindable as Exports
 import Pandora.Pattern.Functor.Distributive as Exports
 import Pandora.Pattern.Functor.Traversable as Exports
 import Pandora.Pattern.Functor.Extractable as Exports
-import Pandora.Pattern.Functor.Pointable as Exports
 import Pandora.Pattern.Functor.Monoidal as Exports
 import Pandora.Pattern.Functor.Semimonoidal as Exports
 import Pandora.Pattern.Functor.Invariant as Exports