+++ CHANGELOG.md
@@ -563,3 +563,4 @@
 * Define `point_` method
 * Use `Monoidal` superclass in `Traversable` instead of `Pointable`
 * Use `Monoidal` superclass in `Monad` instead of `Pointable`
+* Remove `Pointable` typeclass